fix(#13 CodeCleaning): fix clang-tidy issues

+++ types/include/Band.h
@@ -72,7 +72,7 @@ public:
   auto setUpperFrequencyLimit(int newUpperFrequencyLimit) -> void;
   [[nodiscard]] auto getLowerFrequencyLimit() const -> int;
   auto setLowerFrequencyLimit(int newLowerFrequencyLimit) -> void;
-  auto getEffectsSize() -> size_t;
+  auto getEffectsSize() -> int;
   auto getEffectAt(int index) -> haptics::types::Effect&;
   auto addEffect(haptics::types::Effect &newEffect) -> void;
   [[nodiscard]] auto isOverlapping(haptics::types::Effect &effect, const int start, const int stop)
