+++ types/src/Track.cpp
@@ -75,8 +75,8 @@ auto Track::setBodyPartMask(int newBodyPartMask) -> void {
   bodyPartMask = newBodyPartMask;
 }
 
-auto Track::getVerticesSize() -> size_t {
-  return vertices.size();
+auto Track::getVerticesSize() -> int {
+  return static_cast<int>(vertices.size());
 }
 
 auto Track::getVerticeAt(int index) -> int& {
@@ -87,8 +87,8 @@ auto Track::addVertice(int& newVertice) -> void {
   vertices.push_back(newVertice);
 }
 
-auto Track::getBandsSize() -> size_t {
-  return bands.size();
+auto Track::getBandsSize() -> int {
+  return static_cast<int>(bands.size());
 }
 
 auto Track::getBandAt(int index) -> haptics::types::Band& {
